Refactored helper AppTheme to appTheme
